Version 0.2.6
- Fix compiler error for ambiguous requestFrom with 1.5.0-rc.2.

+++ src/LIS3DH.cpp
@@ -389,7 +389,7 @@ bool LIS3DHI2C::readData(uint8_t addr, uint8_t *buf, size_t numBytes) {
 		return false;
 	}
 
-	wire.requestFrom((int)getI2CAddr(), numBytes);
+	wire.requestFrom(getI2CAddr(), numBytes);
 	for(size_t ii = 0; ii < numBytes && wire.available(); ii++) {
 		buf[ii] = wire.read();
 	}
